Finally got the redux state to work
